server test hook invocations during a wait to support fhir requests
