+++ lib/davinci_crd_test_kit/server/v2.0.1/server_urls.rb
@@ -1,0 +1,13 @@
+require_relative '../server_base_urls'
+
+module DaVinciCRDTestKit
+  module V201
+    module ServerURLs
+      include ServerBaseURLs
+
+      def suite_id
+        DaVinciCRDTestKit::V201::CRDServerSuite.id
+      end
+    end
+  end
+end
