+++ lib/davinci_crd_test_kit/server/v2.2.0/server_encounter_start_group.rb
@@ -1,5 +1,5 @@
 require_relative '../../cross_suite/tags'
-require_relative 'verify_request/service_call_test'
+require_relative 'interaction/server_invoke_hook_test'
 require_relative 'verify_request/service_request_required_fields_validation_test'
 require_relative 'verify_request/service_request_optional_fields_validation_test'
 require_relative 'verify_request/service_request_context_validation_test'
@@ -44,7 +44,7 @@ module DaVinciCRDTestKit
       group do
         title 'Make Hook Requests'
 
-        test from: :crd_v220_service_call_test,
+        test from: :crd_v220_server_invoke_hook_test,
              config: {
                inputs: {
                  service_ids: {
