+++ spec/davinci_crd_test_kit/server_invoke_hook_test_spec.rb
@@ -1,36 +1,34 @@
-RSpec.describe DaVinciCRDTestKit::V201::ServiceCallTest do
+RSpec.describe DaVinciCRDTestKit::V201::ServerInvokeHookTest do
   let(:suite_id) { 'crd_server' }
   let(:runnable) do
-    Class.new(DaVinciCRDTestKit::V201::ServiceCallTest) do
+    Class.new(described_class) do
       input :inferno_base_url
     end
   end
   let(:base_url) { 'http://example.com' }
   let(:discovery_url) { 'http://example.com/cds-services' }
   let(:inferno_base_url) { 'http://inferno.com' }
-  let(:service_ids) { 'service_ids' }
+  let(:service_ids) { 'appointment-book-service' }
   let(:service_request_body) do
     json = File.read(File.join(__dir__, '..', 'fixtures', 'appointment_book_hook_request.json'))
     JSON.parse(json)
   end
   let(:service_request_bodies) { [service_request_body].to_json }
-
   let(:encryption_method) { 'ES384' }
+
+  before do
+    allow_any_instance_of(DaVinciCRDTestKit::Jobs::InvokeHook) # hook invocations
+      .to receive(:perform).and_return(nil)
+  end
 
   describe 'when testing a specific hook' do
     before do
       allow_any_instance_of(runnable).to receive(:tested_hook_name).and_return('appointment-book')
     end
 
-    it 'passes when the server returns a 200 HTTP response' do
-      stub_request(:post, "#{discovery_url}/#{service_ids}")
-        .with(
-          body: service_request_body
-        )
-        .to_return(status: 200, body: {}.to_json)
-
+    it 'waits when provided details are sufficient' do
       result = run(runnable, base_url:, inferno_base_url:, service_ids:, encryption_method:, service_request_bodies:)
-      expect(result.result).to eq('pass')
+      expect(result.result).to eq('wait')
     end
 
     it 'skips when the service_ids is not provided' do
@@ -46,18 +44,6 @@ RSpec.describe DaVinciCRDTestKit::V201::ServiceCallTest do
       expect(result.result_message).to match(/Request body not provided/)
     end
 
-    it 'fails when the server does not return a 200 HTTP response' do
-      stub_request(:post, "#{discovery_url}/#{service_ids}")
-        .with(
-          body: service_request_body
-        )
-        .to_return(status: 400, body: {}.to_json)
-
-      result = run(runnable, base_url:, inferno_base_url:, service_ids:, encryption_method:, service_request_bodies:)
-      expect(result.result).to eq('fail')
-      expect(result.result_message).to match(/Unexpected response status: expected 200/)
-    end
-
     it 'fails when the request body is an invalid json' do
       stub_request(:post, "#{discovery_url}/#{service_ids}")
         .with(
@@ -69,30 +55,6 @@ RSpec.describe DaVinciCRDTestKit::V201::ServiceCallTest do
                              service_request_bodies: 'body')
       expect(result.result).to eq('fail')
       expect(result.result_message).to match(/Invalid JSON/)
-    end
-
-    it 'fails when the response body is an invalid json' do
-      stub_request(:post, "#{discovery_url}/#{service_ids}")
-        .with(
-          body: service_request_body
-        )
-        .to_return(status: 200, body: 'response')
-
-      result = run(runnable, base_url:, inferno_base_url:, service_ids:, encryption_method:, service_request_bodies:)
-      expect(result.result).to eq('fail')
-      expect(result.result_message).to match(/Invalid JSON/)
-    end
-
-    it 'makes multiple requests when there are multiple bodies' do
-      hook_invocation = stub_request(:post, "#{discovery_url}/#{service_ids}")
-        .to_return(status: 200, body: {}.to_json)
-
-      multiple_bodies = [service_request_body, service_request_body]
-
-      result = run(runnable, base_url:, inferno_base_url:, service_ids:, encryption_method:,
-                             service_request_bodies: multiple_bodies.to_json)
-      expect(result.result).to eq('pass')
-      expect(hook_invocation).to have_been_made.times(2)
     end
   end
 
@@ -130,14 +92,15 @@ RSpec.describe DaVinciCRDTestKit::V201::ServiceCallTest do
         tags: [DaVinciCRDTestKit::DISCOVERY_TAG]
       )
 
-      stub_request(:post, "#{discovery_url}/appointment-book-service")
-        .to_return(status: 200, body: {}.to_json)
+      expect_any_instance_of(DaVinciCRDTestKit::Jobs::InvokeHook) # rubocop:disable RSpec/StubbedMock
+        .to receive(:perform)
+        .with(anything, anything, "#{discovery_url}/#{service_ids}", anything, anything, anything,
+              DaVinciCRDTestKit::ANY_HOOK_TAG, anything, anything, anything)
+        .and_return(nil)
 
       result = run(runnable, base_url:, inferno_base_url:, service_ids: '', encryption_method:,
                              service_request_bodies:)
-      expect(result.result).to eq('pass')
-
-      expect(session_data_repo.load(test_session_id: test_session.id, name: :invoked_hook)).to eq('appointment-book')
+      expect(result.result).to eq('wait')
     end
   end
 end
